% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ralsaGUIfailsafe.r
\name{ralsaGUIfailsafe}
\alias{ralsaGUIfailsafe}
\title{Start RALSA's Graphical User Interface (GUI) in a failsafe mode}
\usage{
ralsaGUIfailsafe()
}
\description{
If, for any reason, the GUI cannot be started using the regular \code{ralsaGUI()} command, use this instead. The downside of using this function is that the R console will be blocked.
}
