% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ralsaGUI.r
\name{ralsaGUI}
\alias{ralsaGUI}
\title{Start RALSA's Graphical User Interface (GUI)}
\usage{
ralsaGUI()
}
\description{
Starts the GUI. The GUI contains elements for each parameter of each function (data preparation or analysis). The GUI starts as a background job without blocking the console which can be used as usual.
}
